restructure:
1. create MySQLDatabaseUtils in platform, remove db config in app
2. move ProcessingTimeLogInterceptor to platform
3. move TilesConfig to website
