+++ app/rest/src/main/java/sjph/life/app/rest/config/RootApplicationContextConfig.java
@@ -1,11 +1,6 @@
 package sjph.life.app.rest.config;
 
-import javax.sql.DataSource;
-
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
 
 /**
  * @author shaohuiguo
@@ -14,24 +9,4 @@ import org.springframework.jdbc.datasource.DriverManagerDataSource;
 @Configuration
 public class RootApplicationContextConfig {
 
-    /**
-     * @return a {@link DataSource} of MySQL
-     */
-    @Bean
-    public DataSource dataSource() {
-        DriverManagerDataSource driverManagerDataSource = new DriverManagerDataSource();
-        driverManagerDataSource.setDriverClassName("com.mysql.jdbc.Driver");
-        driverManagerDataSource.setUrl("jdbc:mysql://localhost:3306/life_common");
-        driverManagerDataSource.setUsername("root");
-        driverManagerDataSource.setPassword("admin");
-        return driverManagerDataSource;
-    }
-
-    /**
-     * @return a {@link JdbcTemplate} for MySQL {@link DataSource}
-     */
-    @Bean
-    public JdbcTemplate getJdbcTemplate() {
-        return new JdbcTemplate(dataSource());
-    }
 }
