+++ model/src/main/java/sjph/life/model/dao/impl/RelationshipDaoImpl.java
@@ -5,6 +5,7 @@ import java.sql.SQLException;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
@@ -88,6 +89,7 @@ public class RelationshipDaoImpl implements RelationshipDao {
     private final FollowerIdRowMapper followerIdRowMapper       = new FollowerIdRowMapper();
     private final FolloweeIdRowMapper followeeIdRowMapper       = new FolloweeIdRowMapper();
     @Autowired(required = true)
+    @Qualifier("mysqlJdbcTemplate")
     private JdbcTemplate              jdbcTemplate;
 
     @Override
